backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-multipart
httpx
asyncpg
email-validator
google-generativeai